All Grading Requirements Completed
Wrote/Updated .txt files; made it impossible to have x or y
values outside of valid range; Now, however you choose to 
select an area(down and to the left mouse drag vs. up and to 
the right dragging), the proper coordinates will be taken and 
zoomed in on.

All Done!-DK

Signed-off-by: dekuh <dekuh@WINDOWS-A2RHPL0>